requirements.txt
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
fake-useragent==1.5.1
filelock==3.16.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
ipykernel==6.29.5
ipython==8.29.0
itemadapter==0.9.0
itemloaders==1.3.2
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parse==1.20.2
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
Protego==0.3.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==11.1.1
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
requests-toolbelt==1.0.0
schedule==1.2.2
Scrapy==2.11.2
selenium==4.25.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tldextract==5.1.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
w3lib==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
zope.interface==7.1.0
